+++ src/public/commands.ts
@@ -103,6 +103,7 @@ export interface Commands {
 
   switchMode: (mathfield: Mathfield, mode: ParseMode) => boolean;
   insert: (mathfield: Mathfield, s: string, options: InsertOptions) => boolean;
+  insertDecimalSeparator: (mathfield: Mathfield) => boolean;
   typedText: (
     text: string,
     options: {